+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `gent/services-opening-hours` package.
 ### Added
 
 * DMOH-3: Added the Service value object.
+* DMOH-4: Added the Channel value object.
 * DMOH-5: Added the ServiceCollection value object.
 
 [Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop